Formalize Extended namespace, remove Models namespace.
