+++ samples/My.Hr/My.Hr.Api/ImplicitUsings.cs
@@ -6,7 +6,7 @@ global using CoreEx.HealthChecks;
 global using CoreEx.Http;
 global using CoreEx.Json;
 global using CoreEx.RefData;
-global using CoreEx.RefData.Models;
+global using CoreEx.RefData.Extended;
 global using CoreEx.Validation;
 global using CoreEx.WebApis;
 global using Microsoft.AspNetCore.Mvc;
