+++ samples/My.Hr/My.Hr.Business/Models/Gender.cs
@@ -1,8 +1,8 @@
-﻿using CoreEx.RefData;
+﻿using CoreEx.RefData.Extended;
 
 namespace My.Hr.Business.Models;
 
-public class Gender : ReferenceDataBase<Guid, Gender> 
+public class Gender : ReferenceDataBaseEx<Guid, Gender> 
 {
     public static implicit operator Gender?(string? code) => ConvertFromCode(code);
 }
