+++ samples/My.Hr/My.Hr.Infra.Tests/My.Hr.Infra.Tests.csproj
@@ -16,7 +16,7 @@
     <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
     <PackageReference Include="NUnit.Analyzers" Version="3.3.0" />
     <PackageReference Include="coverlet.collector" Version="3.1.2" />
-    <PackageReference Include="UnitTestEx.NUnit" Version="1.0.26" />
+    <PackageReference Include="UnitTestEx.NUnit" Version="1.0.27" />
   </ItemGroup>
 
   <ItemGroup>
