+++ samples/My.Hr/My.Hr.UnitTest/My.Hr.UnitTest.csproj
@@ -24,8 +24,8 @@
     <PackageReference Include="NUnit" Version="3.13.3" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
     <PackageReference Include="coverlet.collector" Version="3.1.0" />
-    <PackageReference Include="UnitTestEx" Version="1.0.26" />
-    <PackageReference Include="UnitTestEx.NUnit" Version="1.0.26" />
+    <PackageReference Include="UnitTestEx" Version="1.0.27" />
+    <PackageReference Include="UnitTestEx.NUnit" Version="1.0.27" />
   </ItemGroup>
 
   <ItemGroup>
