+++ src/CoreEx/RefData/IReferenceData.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/CoreEx
 
 using CoreEx.Entities;
+using CoreEx.RefData.Extended;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
@@ -56,7 +57,7 @@ namespace CoreEx.RefData
         public bool IsValid => true;
 
         /// <summary>
-        /// Overrides the standard <see cref="IsValid"/> check and flags the <see cref="ReferenceDataBase{TId, TSelf}"/> as <b>Invalid</b>.
+        /// Overrides the standard <see cref="IsValid"/> check and flags the <see cref="ReferenceDataBaseEx{TId, TSelf}"/> as <b>Invalid</b>.
         /// </summary>
         /// <remarks>Will result in <see cref="IsActive"/> set to <c>false</c>. Once set to invalid it can not be changed; i.e. there is not an means to set back to valid.</remarks>
         void SetInvalid() { }
@@ -85,7 +86,7 @@ namespace CoreEx.RefData
         public void SetMapping<T>(string name, T? value) where T : IComparable<T?>, IEquatable<T?> => throw new NotImplementedException();
 
         /// <summary>
-        /// Gets a mapping value for the <see cref="ReferenceDataBase{TId, TSelf}"/> for the specified <paramref name="name"/>.
+        /// Gets a mapping value for the <see cref="ReferenceDataBaseEx{TId, TSelf}"/> for the specified <paramref name="name"/>.
         /// </summary>
         /// <typeparam name="T">The value <see cref="Type"/>.</typeparam>
         /// <param name="name">The mapping name.</param>
@@ -93,7 +94,7 @@ namespace CoreEx.RefData
         public T? GetMapping<T>(string name) where T : IComparable<T?>, IEquatable<T?> => default!;
 
         /// <summary>
-        /// Gets a mapping value for the <see cref="ReferenceDataBase{TId, TSelf}"/> for the specified <paramref name="name"/>.
+        /// Gets a mapping value for the <see cref="ReferenceDataBaseEx{TId, TSelf}"/> for the specified <paramref name="name"/>.
         /// </summary>
         /// <typeparam name="T">The value <see cref="Type"/>.</typeparam>
         /// <param name="name">The mapping name.</param>
