+++ tests/CoreEx.Cosmos.Test/CoreEx.Cosmos.Test.csproj
@@ -31,7 +31,7 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="UnitTestEx.NUnit" Version="1.0.26" />
+    <PackageReference Include="UnitTestEx.NUnit" Version="1.0.27" />
   </ItemGroup>
 
   <ItemGroup>
