+++ tests/CoreEx.Test/Framework/Abstractions/Reflection/TypeReflectorTest.cs
@@ -1,6 +1,6 @@
 ﻿using CoreEx.Abstractions.Reflection;
 using CoreEx.Entities;
-using CoreEx.RefData.Models;
+using CoreEx.RefData;
 using NUnit.Framework;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
